requirements.txt
PyYAML
requests
geoip2
